Move the import/export buttons to the Appearance tab #5790
