+++ src/main/java/org/jabref/gui/preferences/AppearanceTab.fxml
@@ -9,6 +9,8 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Tooltip?>
 <fx:root prefWidth="650.0" spacing="10.0" type="VBox" xmlns="http://javafx.com/javafx/11.0.1"
          xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.preferences.AppearanceTabView">
     <fx:define>
@@ -30,4 +32,17 @@
     <RadioButton fx:id="themeLight" text="%Light theme" toggleGroup="$theme"/>
     <RadioButton fx:id="themeDark" text="%Dark theme" toggleGroup="$theme"/>
     <RadioButton fx:id="customTheme" text="%Custom theme" toggleGroup="$theme"/>
+
+    <Label styleClass="sectionHeader" text="%Import custom theme" />
+    <Button maxWidth="Infinity" onAction="#importTheme" text="%Import theme">
+        <tooltip>
+            <Tooltip text="%Import custom CSS file" />
+        </tooltip>
+    </Button>
+    <Button maxWidth="Infinity" onAction="#exportTheme" text="%Export theme">
+        <tooltip>
+            <Tooltip text="%Export theme as CSS" />
+        </tooltip>
+    </Button>
+
 </fx:root>
