+++ src/main/resources/l10n/JabRef_en.properties
@@ -2127,6 +2127,7 @@ Theme\ name=Theme name
 Theme\ changed\ to\ a\ custom\ theme.=Theme changed to a custom theme.
 Import\ theme=Import theme
 Export\ Theme=Export Theme
+Import\ custom\ theme=Import custom theme
 
 Import\ CSS=Import CSS
 Import\ custom\ CSS\ file=Import custom CSS file
